+++ .vscode/launch.json
@@ -14,7 +14,7 @@
       "name": "Web",
       "request": "launch",
       "type": "dart",
-      "args": ["-d", "chrome", "--web-renderer", "skwasm"],
+      "args": ["-d", "chrome", "--web-renderer", "skwasm", "--web-port", "5000"],
     },
   ]
 }
