+++ lib/data.dart
@@ -5,4 +5,6 @@ export "src/data/types.dart";
 export "src/data/user_profile.dart";
 export "src/data/seller_profile.dart";
 export "src/data/product.dart";
+export "src/data/product_sort.dart";
+export "src/data/product_filters.dart";
 export "src/data/review.dart";
