+++ lib/src/data/product_filters.dart
@@ -1,0 +1,9 @@
+import "package:btc_market/data.dart";
+
+class ProductFilters {
+  Set<Category> categories = {};
+  int? minRating;
+  String? searchQuery;
+  int? minPrice;
+  int? maxPrice;
+}
